Add files via upload
Who doesn't like a good glass of wine? The aim of this project was to predict the quality of red wine using different substance parameters present in the wine as features. This is a classic example of supervised learning and I have utilized the traditional ML models to obtain high accuracy.